etl/requirements.txt
pyyaml==5.3.1
backoff==1.11.1
clickhouse-driver==0.2.2
kafka-python==2.0.2
pydantic==1.8.2